paperref/frontend/pyproject.toml
[project]
name = "paperref-frontend"
version = "0.1.0"
description = "PaperRef frontend implemented in Streamlit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit",
    "extra-streamlit-components",
    "pandas",
    "pyjwt",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]